[#161] Seperate file extractor & file upload
